explicit type conversion +

+++ mfc_curve2/frame.cpp
@@ -169,29 +169,29 @@ afx_msg void RightTopFrame::OnPaint()
 		switch (pf.CURVE_STATE)	
 		{
 		case ELLIPS:
-			m_picDC.Ellipse(O.x - step*pf.ca, O.y - step*pf.cb, O.x + step*pf.ca, O.y + step*pf.cb);
+			m_picDC.Ellipse((int)(O.x - step*pf.ca), (int)(O.y - step*pf.cb), (int)(O.x + step*pf.ca), (int)(O.y + step*pf.cb));
 			break;
 		case PARABOLA:
 			if (pf.cp < 0)
 				pf.cp *= -1;	// т.к. существует поворот
 			for (DOUBLE x = O.x; x <= rect.right; ++x)
 			{
-				m_picDC.SetPixel(x, O.y - sqrt(2 * pf.cp*step*(x - O.x)), BLACK_PEN);	// верхн€€
-				m_picDC.SetPixel(x, O.y + sqrt(2 * pf.cp*step*(x - O.x)), BLACK_PEN);	// нижн€€
+				m_picDC.SetPixel((int)x, (int)(O.y - sqrt(2 * pf.cp*step*(x - O.x))), BLACK_PEN);	// верхн€€
+				m_picDC.SetPixel((int)x, (int)(O.y + sqrt(2 * pf.cp*step*(x - O.x))), BLACK_PEN);	// нижн€€
 			}
 			break;
 		case HIPERBOLA:
 			// мучалс€ 2 часа :)
 			for (DOUBLE x = O.x + pf.ca*step; x < rect.right; ++x)
 			{
-				m_picDC.SetPixel(x, ceil(O.y - step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
-				m_picDC.SetPixel(x, ceil(O.y + step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
+				m_picDC.SetPixel((int)x, (int)ceil(O.y - step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
+				m_picDC.SetPixel((int)x, (int)ceil(O.y + step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
 			}
 
 			for (DOUBLE x = O.x - pf.ca*step; x >= rect.left; --x)
 			{
-				m_picDC.SetPixel(x, ceil(O.y - step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
-				m_picDC.SetPixel(x, ceil(O.y + step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
+				m_picDC.SetPixel((int)x, (int)ceil(O.y - step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
+				m_picDC.SetPixel((int)x, (int)ceil(O.y + step*pf.cb*sqrt((x - O.x)*(x - O.x) / (step*step*pf.ca*pf.ca) - 1)), BLACK_PEN);
 			}
 			break;
 		case COINCIDING:
@@ -203,10 +203,10 @@ afx_msg void RightTopFrame::OnPaint()
 			m_picDC.Ellipse(O.x - 3, O.y - 3, O.x + 3, O.y + 3);
 			break;
 		case PARALLEL:
-			m_picDC.MoveTo(O.x - step*pf.ca, 0);
-			m_picDC.LineTo(O.x - step*pf.ca, rect.bottom);
-			m_picDC.MoveTo(O.x + step*pf.ca, 0);
-			m_picDC.LineTo(O.x + step*pf.ca, rect.bottom);
+			m_picDC.MoveTo((int)(O.x - step*pf.ca), 0);
+			m_picDC.LineTo((int)(O.x - step*pf.ca), rect.bottom);
+			m_picDC.MoveTo((int)(O.x + step*pf.ca), 0);
+			m_picDC.LineTo((int)(O.x + step*pf.ca), rect.bottom);
 			break;
 		default:
 			MessageBox(TEXT("”равнение кривой второго пор€дка введено неверно!"),
